+++ src/manage/denyAccessRequest.ts
@@ -20,6 +20,7 @@
 import type { VerifiableCredential } from "@inrupt/solid-client-vc";
 import type { UrlString, WebId } from "@inrupt/solid-client";
 import {
+  CREDENTIAL_TYPE_ACCESS_DENIAL,
   GC_CONSENT_STATUS_DENIED,
   GC_CONSENT_STATUS_EXPLICITLY_GIVEN,
 } from "../constants";
@@ -52,7 +53,7 @@ async function denyAccessRequest(
   const internalOptions = initializeGrantParameters(
     baseAccessVerifiableCredential
   );
-  const requestBody = getGrantBody({
+  const denialBody = getGrantBody({
     resourceOwner,
     access: internalOptions.access,
     requestor: internalOptions.requestor,
@@ -61,14 +62,11 @@ async function denyAccessRequest(
     status: GC_CONSENT_STATUS_EXPLICITLY_GIVEN,
     // eslint-disable-next-line @typescript-eslint/no-explicit-any
   }) as any; // FIXME
-  requestBody.credentialSubject.providedConsent.hasStatus =
+  denialBody.type = [CREDENTIAL_TYPE_ACCESS_DENIAL];
+  denialBody.credentialSubject.providedConsent.hasStatus =
     GC_CONSENT_STATUS_DENIED;
 
-  return issueAccessOrConsentVc(
-    requestBody.credentialSubject.id,
-    requestBody,
-    options
-  );
+  return issueAccessOrConsentVc(resourceOwner, denialBody, options);
 }
 
 export { denyAccessRequest };